Tidying scripts, controllers
